File organisation refactoring
