+++ python_interface/logging.conf
@@ -42,4 +42,4 @@ formatter=customFormatter
 args=(sys.stdout,)
 
 [formatter_customFormatter]
-class=custom_formatter.CustomFormatter
+class=custom.custom_formatter.CustomFormatter
